+++ qualibration_libs/parameters/__init__.py
@@ -1,11 +1,13 @@
 from .common import CommonNodeParameters
-from .experiment import QubitsExperimentNodeParameters, get_qubits
+from .experiment import QubitPairExperimentNodeParameters, QubitsExperimentNodeParameters, get_qubit_pairs, get_qubits
 from .sweep import IdleTimeNodeParameters, get_idle_times_in_clock_cycles
 
 __all__ = [
     "CommonNodeParameters",
     "QubitsExperimentNodeParameters",
+    "QubitPairExperimentNodeParameters",
     "get_qubits",
+    "get_qubit_pairs",
     "IdleTimeNodeParameters",
     "get_idle_times_in_clock_cycles",
 ]