+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+- Add TwoQubitExperimentNodeParameters class
 ### Added
 - New dependencies: lmfit, matplotlib, numpy, qm, quam-builder, qualibrate, scipy, xarray